// src/Models/order.ts
import { item } from './item';

export class order {
  id: string = '';
  order_no: string = '';
  name: string = '';
  order_date: string = '';
  // ItemsOrdered: item[];
  order_total: string = '';
  Tax_total: string = '';
  Grand_total: string = '';
  constructor() // name: string, // order_no: string,
  // date: string,
  // ItemsOrdered: item[]
  // order_total: string,
  // Tax_total: string,
  // Grand_total: string
  {
    // this.order_no = order_no;
    // this.name = name;
    // this.date = date;
    // this.ItemsOrdered = [];
    //   this.order_total = order_total;
    //   this.Tax_total = Tax_total;
    //   this.Grand_total = Grand_total;
    // }
  }
}